+++ lib/packages/wikibase/data-model/src/Entity/PropertyDataTypeLookup.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace Wikibase\DataModel\Entity;
+
+/**
+ * Interface for objects that can find the if of the DataType
+ * for the Property of which the id is given.
+ *
+ * @since 1.0
+ *
+ * @licence GNU GPL v2+
+ * @author Jeroen De Dauw < jeroendedauw@gmail.com >
+ */
+interface PropertyDataTypeLookup {
+
+	/**
+	 * Returns the data type for the Property of which the id is given.
+	 *
+	 * @since 0.4
+	 *
+	 * @param PropertyId $propertyId
+	 *
+	 * @return string
+	 * @throws PropertyNotFoundException
+	 */
+	public function getDataTypeIdForProperty( PropertyId $propertyId );
+
+}