+++ lib/packages/wikibase/data-model/tests/unit/Entity/InMemoryDataTypeLookupTest.php
@@ -1,0 +1,49 @@
+<?php
+
+namespace Wikibase\Test;
+
+use Wikibase\DataModel\Entity\InMemoryDataTypeLookup;
+use Wikibase\DataModel\Entity\PropertyId;
+
+/**
+ * @covers Wikibase\DataModel\Entity\InMemoryDataTypeLookup
+ * @uses Wikibase\DataModel\Entity\PropertyId
+ * @uses Wikibase\DataModel\Entity\PropertyNotFoundException
+ *
+ * @licence GNU GPL v2+
+ * @author Jeroen De Dauw < jeroendedauw@gmail.com >
+ */
+class InMemoryDataTypeLookupTest extends \PHPUnit_Framework_TestCase {
+
+	public function testGetDataTypeForPropertyThatIsNotSet() {
+		$lookup = new InMemoryDataTypeLookup();
+
+		$this->setExpectedException( 'Wikibase\DataModel\Entity\PropertyNotFoundException' );
+
+		$lookup->getDataTypeIdForProperty( new PropertyId( 'p7201010' ) );
+	}
+
+	public function testSetAndGetDataType() {
+		$propertyId = new PropertyId( 'p7201010' );
+
+		$stringTypeId = 'string-datatype';
+		$intTypeId = 'integer';
+
+		$lookup = new InMemoryDataTypeLookup();
+		$lookup->setDataTypeForProperty( $propertyId, $stringTypeId );
+
+		$actual = $lookup->getDataTypeIdForProperty( $propertyId );
+
+		$this->assertInternalType( 'string', $actual );
+
+		$this->assertEquals( $stringTypeId, $actual );
+
+		$lookup->setDataTypeForProperty( $propertyId, $intTypeId );
+
+		$actual = $lookup->getDataTypeIdForProperty( $propertyId );
+
+		$this->assertNotEquals( $stringTypeId, $actual );
+		$this->assertEquals( $intTypeId, $actual );
+	}
+
+}